option_trades/requirements.txt
quixstreams
python-dotenv
websockets
quixplus
pandas